iroh_test::logging
Function setup_multithreaded

pub fn setup_multithreaded()

The first call to this function will install a global logger.

The logger uses the RUST_LOG environment variable to decide on what level to log anything, which is set by our CI. When running the tests with nextest the log output will be captured for just the executing test.

Logs to stdout since the assertion messages are logged on stderr by default.