Struct iroh_test::logging::TestWriter

struct TestWriter;

A tracing writer that interacts well with test output capture.

Using this writer will make sure that the output is captured normally and only printed when the test fails. See setup to actually use this.